[render] Split up the concept of an object and a mesh
